+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed the Python and R Packages views incorrectly stating the default package
   files were missing even when they were present (#2882, #2884)
+- Fixed an issue where the package file checkbox in the Project Files view
+  would occassionally be unchecked even when the file was included (#2793)
 
 ### Changed